Split up household_net_income file per variable
