+++ policyengine_us/variables/household/income/household/household_count_people.py
@@ -1,0 +1,12 @@
+from policyengine_us.model_api import *
+
+
+class household_count_people(Variable):
+    value_type = int
+    entity = Household
+    label = "Number of people"
+    definition_period = YEAR
+    unit = "person"
+
+    def formula(household, period, parameters):
+        return household.nb_persons()
