+++ policyengine_us/variables/household/income/household/person_in_poverty.py
@@ -1,0 +1,12 @@
+from policyengine_us.model_api import *
+
+
+class person_in_poverty(Variable):
+    label = "person in poverty"
+    documentation = "Whether person is in poverty"
+    entity = Person
+    definition_period = YEAR
+    value_type = bool
+
+    def formula(person, period, parameters):
+        return person.household("in_poverty", period)
