+++ policyengine_us/variables/household/income/household/poverty_gap.py
@@ -1,0 +1,10 @@
+from policyengine_us.model_api import *
+
+
+class poverty_gap(Variable):
+    label = "poverty gap"
+    documentation = "Poverty gap"
+    entity = Household
+    definition_period = YEAR
+    value_type = float
+    unit = USD
